LayoutTests/imported/w3c/web-platform-tests/IndexedDB/idbobjectstore_createIndex6-event_order-expected.txt


PASS IDBObjectStore.createIndex() - event order when unique constraint is triggered
